+++ src/Controller/UsersController.php
@@ -15,6 +15,7 @@ use CakeDC\Users\Controller\AppController;
 use CakeDC\Users\Controller\Component\UsersAuthComponent;
 use CakeDC\Users\Controller\Traits\LoginTrait;
 use CakeDC\Users\Controller\Traits\ProfileTrait;
+use CakeDC\Users\Controller\Traits\ReCaptchaTrait;
 use CakeDC\Users\Controller\Traits\RegisterTrait;
 use CakeDC\Users\Controller\Traits\SimpleCrudTrait;
 use CakeDC\Users\Controller\Traits\SocialTrait;
@@ -31,18 +32,8 @@ class UsersController extends AppController
 {
     use LoginTrait;
     use ProfileTrait;
+    use ReCaptchaTrait;
     use RegisterTrait;
     use SimpleCrudTrait;
     use SocialTrait;
-
-    /**
-     * Override loadModel to load specific users table
-     * @param string $modelClass model class
-     * @param string $type type
-     * @return Table
-     */
-    public function loadModel($modelClass = null, $type = 'Table')
-    {
-        return parent::loadModel(Configure::read('Users.table'));
-    }
 }