+++ src/Model/Behavior/SocialAccountBehavior.php
@@ -62,6 +62,7 @@ class SocialAccountBehavior extends Behavior
         if (empty($user)) {
             return true;
         }
+
         return $this->sendSocialValidationEmail($entity, $user);
     }
 
@@ -98,10 +99,10 @@ class SocialAccountBehavior extends Behavior
 
         if (!empty($socialAccount) && $socialAccount->token === $token) {
             if ($socialAccount->active) {
-                throw new AccountAlreadyActiveException(__d('Users', "Account already validated"));
+                throw new AccountAlreadyActiveException(__d('CakeDC/Users', "Account already validated"));
             }
         } else {
-            throw new RecordNotFoundException(__d('Users', "Account not found for the given token and email."));
+            throw new RecordNotFoundException(__d('CakeDC/Users', "Account not found for the given token and email."));
         }
 
         return $this->_activateAccount($socialAccount);
@@ -125,10 +126,10 @@ class SocialAccountBehavior extends Behavior
 
         if (!empty($socialAccount)) {
             if ($socialAccount->active) {
-                throw new AccountAlreadyActiveException(__d('Users', "Account already validated"));
+                throw new AccountAlreadyActiveException(__d('CakeDC/Users', "Account already validated"));
             }
         } else {
-            throw new RecordNotFoundException(__d('Users', "Account not found for the given token and email."));
+            throw new RecordNotFoundException(__d('CakeDC/Users', "Account not found for the given token and email."));
         }
 
         return $this->sendSocialValidationEmail($socialAccount, $socialAccount->user);
@@ -144,6 +145,7 @@ class SocialAccountBehavior extends Behavior
     {
         $socialAccount->active = true;
         $result = $this->_table->save($socialAccount);
+
         return $result;
     }
 }