+++ src/Model/Table/SocialAccountsTable.php
@@ -113,6 +113,7 @@ class SocialAccountsTable extends Table
     public function buildRules(RulesChecker $rules)
     {
         $rules->add($rules->existsIn(['user_id'], 'Users'));
+
         return $rules;
     }
 }