+++ src/Shell/UsersShell.php
@@ -51,21 +51,22 @@ class UsersShell extends Shell
     public function getOptionParser()
     {
         $parser = parent::getOptionParser();
-        $parser->description(__d('Users', 'Utilities for CakeDC Users Plugin'))
-            ->addSubcommand('activateUser')->description(__d('Users', 'Activate an specific user'))
-            ->addSubcommand('addSuperuser')->description(__d('Users', 'Add a new superadmin user for testing purposes'))
-            ->addSubcommand('addUser')->description(__d('Users', 'Add a new user'))
-            ->addSubcommand('changeRole')->description(__d('Users', 'Change the role for an specific user'))
-            ->addSubcommand('deactivateUser')->description(__d('Users', 'Deactivate an specific user'))
-            ->addSubcommand('deleteUser')->description(__d('Users', 'Delete an specific user'))
-            ->addSubcommand('passwordEmail')->description(__d('Users', 'Reset the password via email'))
-            ->addSubcommand('resetAllPasswords')->description(__d('Users', 'Reset the password for all users'))
-            ->addSubcommand('resetPassword')->description(__d('Users', 'Reset the password for an specific user'))
+        $parser->description(__d('CakeDC/Users', 'Utilities for CakeDC Users Plugin'))
+            ->addSubcommand('activateUser')->description(__d('CakeDC/Users', 'Activate an specific user'))
+            ->addSubcommand('addSuperuser')->description(__d('CakeDC/Users', 'Add a new superadmin user for testing purposes'))
+            ->addSubcommand('addUser')->description(__d('CakeDC/Users', 'Add a new user'))
+            ->addSubcommand('changeRole')->description(__d('CakeDC/Users', 'Change the role for an specific user'))
+            ->addSubcommand('deactivateUser')->description(__d('CakeDC/Users', 'Deactivate an specific user'))
+            ->addSubcommand('deleteUser')->description(__d('CakeDC/Users', 'Delete an specific user'))
+            ->addSubcommand('passwordEmail')->description(__d('CakeDC/Users', 'Reset the password via email'))
+            ->addSubcommand('resetAllPasswords')->description(__d('CakeDC/Users', 'Reset the password for all users'))
+            ->addSubcommand('resetPassword')->description(__d('CakeDC/Users', 'Reset the password for an specific user'))
             ->addOptions([
                 'username' => ['short' => 'u', 'help' => 'The username for the new user'],
                 'password' => ['short' => 'p', 'help' => 'The password for the new user'],
                 'email' => ['short' => 'e', 'help' => 'The email for the new user']
             ]);
+
         return $parser;
     }
 
@@ -94,11 +95,11 @@ class UsersShell extends Shell
         $userEntity->is_superuser = true;
         $userEntity->role = 'user';
         $savedUser = $this->Users->save($userEntity);
-        $this->out(__d('Users', 'User added:'));
-        $this->out(__d('Users', 'Id: {0}', $savedUser->id));
-        $this->out(__d('Users', 'Username: {0}', $username));
-        $this->out(__d('Users', 'Email: {0}', $savedUser->email));
-        $this->out(__d('Users', 'Password: {0}', $password));
+        $this->out(__d('CakeDC/Users', 'User added:'));
+        $this->out(__d('CakeDC/Users', 'Id: {0}', $savedUser->id));
+        $this->out(__d('CakeDC/Users', 'Username: {0}', $username));
+        $this->out(__d('CakeDC/Users', 'Email: {0}', $savedUser->email));
+        $this->out(__d('CakeDC/Users', 'Password: {0}', $password));
     }
 
     /**
@@ -122,16 +123,16 @@ class UsersShell extends Shell
         $userEntity->role = 'superuser';
         $savedUser = $this->Users->save($userEntity);
         if (!empty($savedUser)) {
-            $this->out(__d('Users', 'Superuser added:'));
-            $this->out(__d('Users', 'Id: {0}', $savedUser->id));
-            $this->out(__d('Users', 'Username: {0}', $username));
-            $this->out(__d('Users', 'Email: {0}', $savedUser->email));
-            $this->out(__d('Users', 'Password: {0}', $password));
+            $this->out(__d('CakeDC/Users', 'Superuser added:'));
+            $this->out(__d('CakeDC/Users', 'Id: {0}', $savedUser->id));
+            $this->out(__d('CakeDC/Users', 'Username: {0}', $username));
+            $this->out(__d('CakeDC/Users', 'Email: {0}', $savedUser->email));
+            $this->out(__d('CakeDC/Users', 'Password: {0}', $password));
         } else {
-            $this->out(__d('Users', 'Superuser could not be added:'));
+            $this->out(__d('CakeDC/Users', 'Superuser could not be added:'));
 
             collection($userEntity->errors())->each(function ($error, $field) {
-                $this->out(__d('Users', 'Field: {0} Error: {1}', $field, implode(',', $error)));
+                $this->out(__d('CakeDC/Users', 'Field: {0} Error: {1}', $field, implode(',', $error)));
             });
         }
     }
@@ -149,12 +150,12 @@ class UsersShell extends Shell
     {
         $password = Hash::get($this->args, 0);
         if (empty($password)) {
-            $this->error(__d('Users', 'Please enter a password.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a password.'));
         }
         $hashedPassword = $this->_generatedHashedPassword($password);
         $this->Users->updateAll(['password' => $hashedPassword], ['id IS NOT NULL']);
-        $this->out(__d('Users', 'Password changed for all users'));
-        $this->out(__d('Users', 'New password: {0}', $password));
+        $this->out(__d('CakeDC/Users', 'Password changed for all users'));
+        $this->out(__d('CakeDC/Users', 'New password: {0}', $password));
     }
 
     /**
@@ -172,17 +173,17 @@ class UsersShell extends Shell
         $username = Hash::get($this->args, 0);
         $password = Hash::get($this->args, 1);
         if (empty($username)) {
-            $this->error(__d('Users', 'Please enter a username.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a username.'));
         }
         if (empty($password)) {
-            $this->error(__d('Users', 'Please enter a password.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a password.'));
         }
         $data = [
             'password' => $password
         ];
         $this->_updateUser($username, $data);
-        $this->out(__d('Users', 'Password changed for user: {0}', $username));
-        $this->out(__d('Users', 'New password: {0}', $password));
+        $this->out(__d('CakeDC/Users', 'Password changed for user: {0}', $username));
+        $this->out(__d('CakeDC/Users', 'New password: {0}', $password));
     }
 
     /**
@@ -200,17 +201,17 @@ class UsersShell extends Shell
         $username = Hash::get($this->args, 0);
         $role = Hash::get($this->args, 1);
         if (empty($username)) {
-            $this->error(__d('Users', 'Please enter a username.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a username.'));
         }
         if (empty($role)) {
-            $this->error(__d('Users', 'Please enter a role.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a role.'));
         }
         $data = [
             'role' => $role
         ];
         $savedUser = $this->_updateUser($username, $data);
-        $this->out(__d('Users', 'Role changed for user: {0}', $username));
-        $this->out(__d('Users', 'New role: {0}', $savedUser->role));
+        $this->out(__d('CakeDC/Users', 'Role changed for user: {0}', $username));
+        $this->out(__d('CakeDC/Users', 'New role: {0}', $savedUser->role));
     }
 
     /**
@@ -225,7 +226,7 @@ class UsersShell extends Shell
     public function activateUser()
     {
         $user = $this->_changeUserActive(true);
-        $this->out(__d('Users', 'User was activated: {0}', $user->username));
+        $this->out(__d('CakeDC/Users', 'User was activated: {0}', $user->username));
     }
 
     /**
@@ -240,7 +241,7 @@ class UsersShell extends Shell
     public function deactivateUser()
     {
         $user = $this->_changeUserActive(false);
-        $this->out(__d('Users', 'User was de-activated: {0}', $user->username));
+        $this->out(__d('CakeDC/Users', 'User was de-activated: {0}', $user->username));
     }
 
     /**
@@ -252,7 +253,7 @@ class UsersShell extends Shell
     {
         $reference = Hash::get($this->args, 0);
         if (empty($reference)) {
-            $this->error(__d('Users', 'Please enter a username or email.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a username or email.'));
         }
         $resetUser = $this->Users->resetToken($reference, [
             'expiration' => Configure::read('Users.Token.expiration'),
@@ -260,10 +261,10 @@ class UsersShell extends Shell
             'sendEmail' => true,
         ]);
         if ($resetUser) {
-            $msg = __d('Users', 'Please ask the user to check the email to continue with password reset process');
+            $msg = __d('CakeDC/Users', 'Please ask the user to check the email to continue with password reset process');
             $this->out($msg);
         } else {
-            $msg = __d('Users', 'The password token could not be generated. Please try again');
+            $msg = __d('CakeDC/Users', 'The password token could not be generated. Please try again');
             $this->error($msg);
         }
     }
@@ -278,11 +279,12 @@ class UsersShell extends Shell
     {
         $username = Hash::get($this->args, 0);
         if (empty($username)) {
-            $this->error(__d('Users', 'Please enter a username.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a username.'));
         }
         $data = [
             'active' => $active
         ];
+
         return $this->_updateUser($username, $data);
     }
 
@@ -295,9 +297,9 @@ class UsersShell extends Shell
      */
     protected function _updateUser($username, $data)
     {
-        $user = $this->Users->find()->where(['Users.username' => $username])->first();
+        $user = $this->Users->find()->where(['username' => $username])->first();
         if (empty($user)) {
-            $this->error(__d('Users', 'The user was not found.'));
+            $this->error(__d('CakeDC/Users', 'The user was not found.'));
         }
         $user = $this->Users->patchEntity($user, $data);
         collection($data)->filter(function ($value, $field) use ($user) {
@@ -306,6 +308,7 @@ class UsersShell extends Shell
             $user->{$field} = $value;
         });
         $savedUser = $this->Users->save($user);
+
         return $savedUser;
     }
 
@@ -318,16 +321,17 @@ class UsersShell extends Shell
     {
         $username = Hash::get($this->args, 0);
         if (empty($username)) {
-            $this->error(__d('Users', 'Please enter a username.'));
+            $this->error(__d('CakeDC/Users', 'Please enter a username.'));
         }
-        $user = $this->Users->find()->where(['Users.username' => $username])->first();
-        $deleteAccounts = $this->Users->SocialAccounts->deleteAll(['user_id' => $user->id]);
+        $user = $this->Users->find()->where(['username' => $username])->first();
+        if (isset($this->Users->SocialAccounts)) {
+            $this->Users->SocialAccounts->deleteAll(['user_id' => $user->id]);
+        }
         $deleteUser = $this->Users->delete($user);
-        if ($deleteAccounts && $deleteUser) {
-            $this->out(__d('Users', 'The user {0} was deleted successfully', $username));
-        } else {
-            $this->error(__d('Users', 'The user {0} was not deleted. Please try again', $username));
+        if (!$deleteUser) {
+            $this->error(__d('CakeDC/Users', 'The user {0} was not deleted. Please try again', $username));
         }
+        $this->out(__d('CakeDC/Users', 'The user {0} was deleted successfully', $username));
     }
 
     /**